refactor(native): split runtime render modules and render tests by subdomain
