+++ src/native/engine/engine.c
@@ -17,7 +17,7 @@
 #include "runtime/input/move_target_service.h"
 #include "runtime/player/player_runtime_service.h"
 #include "runtime/player/player_gateway_abi.h"
-#include "runtime/render/render_material.h"
+#include "runtime/render/material/render_material.h"
 #include "runtime/player/player_motion.h"
 #include "runtime/player/player_motion_host.h"
 #include "runtime/player/player_api.h"
@@ -31,7 +31,7 @@
 #include "runtime/abi/physics_abi.h"
 #include "runtime/input/input_abi.h"
 #include "runtime/resource/resource_domain.h"
-#include "runtime/render/render_submit.h"
+#include "runtime/render/submit/render_submit.h"
 #include "runtime/terrain/terrain_generation.h"
 #include "runtime/terrain/terrain_abi.h"
 #include "runtime/terrain/terrain_runtime.h"
