+++ src/native/engine/runtime/engine/application_service_ports.c
@@ -2,8 +2,8 @@
 
 #include "../input/move_target_service.h"
 #include "../player/player_runtime_service.h"
-#include "../render/render_material.h"
-#include "../render/render_submit.h"
+#include "../render/material/render_material.h"
+#include "../render/submit/render_submit.h"
 #include "../terrain/terrain_height.h"
 #include "../terrain/terrain_host.h"
 #include "../terrain/terrain_mutation.h"
