+++ src/native/engine/runtime/render/material/render_material.h
@@ -1,8 +1,8 @@
 #ifndef BANANA_ENGINE_RUNTIME_RENDER_MATERIAL_H
 #define BANANA_ENGINE_RUNTIME_RENDER_MATERIAL_H
 
-#include "../../render/material.h"
-#include "../../world/entity.h"
+#include "../../../render/material.h"
+#include "../../../world/entity.h"
 
 #ifdef __cplusplus
 extern "C"
